requirements.txt
indeed
Py-StackExchange
ntlk
bs4
boto3
pyyaml